+++ Example/ReactiveDataDisplayManager/Base.lproj/Main.storyboard
@@ -433,6 +433,7 @@
                         <segue destination="GZX-q5-5fd" kind="show" identifier="sizableCollection" id="kbc-gq-BdS"/>
                         <segue destination="pGI-i3-EWf" kind="show" identifier="foldableCollection" id="xiU-it-Uz9"/>
                         <segue destination="cOD-zS-jLq" kind="show" identifier="itemTitleCollection" id="Tuv-8l-eFa"/>
+                        <segue destination="NJ8-lG-GQz" kind="show" identifier="compositionalCollection" id="F32-3a-8Fb"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="tMd-8n-CXM" userLabel="First Responder" sceneMemberID="firstResponder"/>
@@ -785,6 +786,44 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="jRY-Cy-okp" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="3394" y="2240"/>
+        </scene>
+        <!--Collection Compositional View Controller-->
+        <scene sceneID="LhE-pz-dzt">
+            <objects>
+                <viewController id="NJ8-lG-GQz" customClass="CollectionCompositionalViewController" customModule="ReactiveDataDisplayManagerExample" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="wVo-RW-qbI">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="prototypes" translatesAutoresizingMaskIntoConstraints="NO" id="iGK-Ib-Tmb">
+                                <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                                <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                                <collectionViewFlowLayout key="collectionViewLayout" automaticEstimatedItemSize="YES" minimumLineSpacing="10" minimumInteritemSpacing="10" id="OkW-TU-ARv">
+                                    <size key="itemSize" width="128" height="128"/>
+                                    <size key="headerReferenceSize" width="0.0" height="0.0"/>
+                                    <size key="footerReferenceSize" width="0.0" height="0.0"/>
+                                    <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
+                                </collectionViewFlowLayout>
+                                <cells/>
+                            </collectionView>
+                        </subviews>
+                        <viewLayoutGuide key="safeArea" id="InC-B9-tmm"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                        <constraints>
+                            <constraint firstItem="InC-B9-tmm" firstAttribute="bottom" secondItem="iGK-Ib-Tmb" secondAttribute="bottom" id="OJk-aZ-WOl"/>
+                            <constraint firstItem="iGK-Ib-Tmb" firstAttribute="top" secondItem="InC-B9-tmm" secondAttribute="top" id="bLs-an-OGI"/>
+                            <constraint firstItem="InC-B9-tmm" firstAttribute="trailing" secondItem="iGK-Ib-Tmb" secondAttribute="trailing" id="fmk-oF-six"/>
+                            <constraint firstItem="iGK-Ib-Tmb" firstAttribute="leading" secondItem="InC-B9-tmm" secondAttribute="leading" id="ytt-uE-5J9"/>
+                        </constraints>
+                    </view>
+                    <navigationItem key="navigationItem" id="kg2-qy-hN8"/>
+                    <connections>
+                        <outlet property="collectionView" destination="iGK-Ib-Tmb" id="Xx3-OE-H6V"/>
+                    </connections>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="suE-ac-67B" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="4185" y="2258"/>
         </scene>
         <!--Navigation Controller-->
         <scene sceneID="7SR-bl-OnR">
