+++ Example/ReactiveDataDisplayManager/Collection/MainCollectionViewController/MainCollectionViewController.swift
@@ -22,6 +22,7 @@ final class MainCollectionViewController: UIViewController {
         case sizableCollection
         case foldableCollection
         case itemTitleCollection
+        case compositionalCollection
     }
 
     // MARK: - Constants
@@ -35,7 +36,8 @@ final class MainCollectionViewController: UIViewController {
             ("Horizontal image collection", .imageHorizontalCollection),
             ("Sizable collection", .sizableCollection),
             ("Foldable collection", .foldableCollection),
-            ("Collection with item index titles", .itemTitleCollection)
+            ("Collection with item index titles", .itemTitleCollection),
+            ("Collection with compositional layout", .compositionalCollection)
         ]
     }
